backend/requirements.txt
fastapi
uvicorn
sqlmodel
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
psycopg2-binary
email-validator
